move single-dojo code from dojos.py to dojo.py
