+++ dojo_theme/templates/admin_dojos.html
@@ -8,7 +8,7 @@
 </div>
 <div class="container">
     {% for dojo in dojos %}
-    <h2><a href="{{ url_for('pwncollege_dojos.view_dojo', dojo=dojo.reference_id) }}">{{ dojo.name }}</a></h2>
+    <h2><a href="{{ url_for('pwncollege_dojo.view_dojo', dojo=dojo.reference_id) }}">{{ dojo.name }}</a></h2>
     <b>Reference ID: </b><code>{{ dojo.reference_id }}</code>
     <br>
     <b>Dojo ID: </b><code>{{ dojo.hex_dojo_id }}</code>
